+++ src/test/kotlin/io/unthrottled/amii/core/personality/emotions/PositiveEmotionCoreTests.kt
@@ -23,7 +23,7 @@ class PositiveEmotionCoreTests {
       mockRandom
     )
 
-    every { mockRandom.nextInt(1, 100) } returns 50
+    every { mockRandom.nextLong(1, 100) } returns 50
     every { mockRandom.nextInt(2) } returns 1
 
     listOf(
